Add the framework 'Mother'
Import the project 'Mother' as a development framework, and make the
approriate changes to Bishop's build scripts to include this source.

'Mother' is a framework that interacts with the system on which our app
is running, and performs the actual work we require.

This commit also bumps the minimum required Android SDK version to '19'.